Added & configured DMX array with extraction and loopback from/to fixtures
The new DMX variable  is now the DMX array. This is the DMX you should
send out. (Do not use as an input). From now on, if you want to change
DMX data in a prime fixture (the ones that are in the fixtures array)
you have to set its DMXChanged to true. (setDimmer(val) and
receiveDMX(vals[]) do this by default)