Make FixtureArray use TreeMap instead of the idLookupTable mess
The TreeMap implementation is a lot cleaner and takes care of a lot of the issues we were having with the idLookupTable and ArrayList -system.
I chose TreeMap, because of its efficient (O(log(n))) key-finding, and it is actually better, in this case, to have the Map always sorted.